Extract StructureOrchestrator; DFTOrchestrator delegates structure step
Structure generation is an engine-agnostic, structure-class-aware step, so it
no longer belongs inside the DFT orchestrator. Extract the generate→validate→
refine loop (plus structure-class skill loading, validation printing, key
auto-discovery, and the run log) into a new StructureOrchestrator. Its control
flow is deterministic Python; the LLM is used only within bounded steps
(generation, validation), each returning a structured result — a true
orchestrator, not an LLM-driven chat loop.

DFTOrchestrator now composes a StructureOrchestrator and delegates Step 1 of
run_complete_workflow to it (reusing its resolved credentials + run log); it
carries no structure-building methods. build_structure(request,
structure_class='crystal') moves to StructureOrchestrator, where it belongs.

Renames the structure axis from 'scale' to 'structure_class' (crystal /
molecular / condensed / biomolecular) to disambiguate from the existing
simulation-scale axis (periodic_dft / molecular_dft / molecular_dynamics /
mlip) used by SimulationRouter.

Smoke-tested: StructureOrchestrator.build_structure (Si, crystal) 2/2 — skill
loads, validation passes, POSCAR only, no DFT inputs; DFTOrchestrator composes
+ shares creds/log (non-LLM wiring check).
