+++ scilink/agents/sim_agents/__init__.py
@@ -12,6 +12,7 @@ from .force_field_agent import ForceFieldAgent
 from .lammps_analysis import LAMMPSAnalysisAgent
 from .lammps_analysis_updater import LAMMPSAnalysisUpdater
 from .lammps_orchestrator import LAMMPSOrchestrator
+from .structure_orchestrator import StructureOrchestrator
 from .dft_orchestrator import DFTOrchestrator
 from .simulation_orchestrator import SimulationOrchestratorAgent, SimulationMode
 from .simulation_router import SimulationRouter, discover_scale_agents
